requirements.txt
appdirs==1.4.4
babelfish==0.5.5
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
decorator==4.4.2
dogpile.cache==0.9.2
enzyme==0.4.1
guessit==3.1.1
idna==2.9
pathtools==0.1.2
pbr==5.4.5
pysrt==1.1.2
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
rarfile==3.1
rebulk==2.0.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
stevedore==1.32.0
subliminal==2.1.0
urllib3==1.25.9
watchdog==0.10.2
